Changed pretty much everything again.
Realized I have to define things globally and stuff.

Made it so it'll stop the previous song before playing another, fixing the whole "layering thousands of songs over eachother until we run out of RAM" thing.

Made a lot of stuff a lot more generic in preparation for having it done via the code instead of me.